Updates and styles register page
