+++ src/pages/Register/RegisterForm/index.styles.js
@@ -1,0 +1,4 @@
+import styled from "styled-components";
+import { BaseForm } from "../../../components/forms/BaseForm/index.styles";
+
+export const RegisterForm = styled(BaseForm)``;
