+++ src/pages/Register/index.styles.js
@@ -1,0 +1,4 @@
+import styled from "styled-components";
+import { LoginWrapper } from "../Login/index.styles";
+
+export const RegisterWrapper = styled(LoginWrapper)``;
